requirements.txt
aioitertools==0.13.0
aiosqlite==0.22.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
arize-phoenix==12.25.1
arize-phoenix-client==1.27.1
arize-phoenix-evals==2.7.1
arize-phoenix-otel==0.14.0
asttokens==3.0.1
Authlib==1.6.6
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
dotenv==0.9.9
duckdb==1.4.3
email-validator==2.3.0
executing==2.2.1
fastapi==0.125.0
fonttools==4.61.0
googleapis-common-protos==1.72.0
graphql-core==3.2.7
greenlet==3.3.0
grpc-interceptor==0.15.4
grpcio==1.76.0
h11==0.16.0
helper==2.5.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.3
jsonpath-ng==1.7.0
kiwisolver==1.4.9
ldap3==2.9.1
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
numpy==2.3.5
openai==2.11.0
openinference-instrumentation==0.1.42
openinference-instrumentation-openai==0.1.41
openinference-semantic-conventions==0.1.25
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
ply==3.11
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==6.33.2
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
pystache==0.6.8
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
sqlean.py==3.50.4.5
stack-data==0.6.3
starlette==0.50.0
strawberry-graphql==0.270.1
threadpoolctl==3.6.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
uvicorn==0.38.0
wcwidth==0.2.14
wrapt==1.17.3
zipp==3.23.0
